requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
supabase>=2.0.0
lxml>=4.9.0
cloudscraper>=1.2.71
openai>=1.0.0
tiktoken>=0.5.0
flask>=3.0.0
flask-cors>=4.0.0
